fix deposit

+++ contracts/libraries/StructureData.sol
@@ -9,8 +9,8 @@ library StructureData {
     struct OptionParameters {
          uint256 quota;  
          uint8 pricePrecision;
-         int8 strikePriceRatio;  // take, 10% is represented as 1000, precision is 4
-         uint8 interestRate; //take, 0.01% is represented as 1, precision is 4
+         int16 strikePriceRatio;  // take, 10% is represented as 1000, precision is 4
+         uint16 interestRate; //take, 0.01% is represented as 1, precision is 4
      }
 
     struct OptionState {
@@ -18,7 +18,7 @@ library StructureData {
          uint256 totalAmount; 
          uint256 strikePrice;
          uint256 underlyingPrice;
-         uint8 interestRate; //take, 0.01% is represented as 1, precision is 4
+         uint16 interestRate; //take, 0.01% is represented as 1, precision is 4
          uint8 pricePrecision; 
          bool executed; 
     }
